Moved micrometer-1.10.x-plugin to apm-toolkit-activation
