+++ apm-sniffer/apm-toolkit-activation/pom.xml
@@ -33,6 +33,7 @@
         <module>apm-toolkit-opentracing-activation</module>
         <module>apm-toolkit-trace-activation</module>
         <module>apm-toolkit-meter-activation</module>
+        <module>apm-toolkit-micrometer-activation</module>
         <module>apm-toolkit-webflux-activation</module>
         <module>apm-toolkit-logging-common</module>
     </modules>
